- Main.cs 		- Added comments 	- BaseCreations.cs 		- Removed unneccesary comments 		- Added comments to BaseCreations.cs 		- Removed unneccesary variable '_label' 	- JobCreation.cs 		- Moved JobCreation to it's own cs file 		- Added comments 	- RepCreation.cs 		- Moved RepCreation to it's own cs file 		- Added comments 	- BoardCreation.cs 		- Moved BoardCreation to it's own cs file 		- Added Comments 	- BaseList.cs 		- Added comments 		- Moved to namespace Job_application_Database.Classes 	- JobsList.cs 		- Renamed JobsListWindow class to JobsList 		- Moved JobsList to it's own cs file 		- Added comments 	- RepsList.cs 		- Renamed RepsListWindow class to RepsList 		- Moved RepsList to it's own cs file 		- Added comments 	- BaseInfo.cs 		- Moved to namespace Job_application_Database.Classes 		- Added comments 		- Removed SortByNameAcending method 	- Board.cs 		- Moved Board to it's own cs file 		- Added comments 	- Rep.cs 		- Moved Rep to it's own cs file 		- Added comments 	- Job.cs 		- Moved Job to it's own cs file 		- Added comments
